Track CirrusScan API consolidation
CirrusScan now uses a single production API (hosted in C3PO, if anybody
cares) instead of separate organization-centric API instances. Simplify
logic to simply use the new common API endpoint always instead of doing
an organization-specific lookup.

I believe the unit tests have been corrected appropriately. The
`org_type` argument to `getAwsUrl()` now is unreferenced, but I left it
alone in order to minimize the footprint of this change.

This change is a blocker for retirement of the old endpoints.